+++ lib/ts/recipe/emailverification/recipeImplementation.ts
@@ -1,6 +1,7 @@
 import { RecipeInterface, NormalisedConfig } from "./types";
 import { NormalisedAppInfo } from "../../types";
 import Querier from "../../querier";
+import Session from "../session/recipe";
 
 export default class RecipeImplementation implements RecipeInterface {
     querier: Querier;
@@ -44,7 +45,8 @@ export default class RecipeImplementation implements RecipeInterface {
         const response: SendVerifyEmailAPIResponse = await this.querier.post(
             "/user/email/verify/token",
             {},
-            (context) => {
+            async (context) => {
+                context = await Session.getInstanceOrThrow().recipeImpl.attachSessionToRequest(context);
                 return input.config.preAPIHook({
                     ...context,
                     action: "SEND_VERIFY_EMAIL",
@@ -66,7 +68,8 @@ export default class RecipeImplementation implements RecipeInterface {
             "/user/email/verify",
             {},
             undefined,
-            (context) => {
+            async (context) => {
+                context = await Session.getInstanceOrThrow().recipeImpl.attachSessionToRequest(context);
                 return input.config.preAPIHook({
                     ...context,
                     action: "IS_EMAIL_VERIFIED",