+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Changes `redirectToAuth` to take an object and adds `redirectBack` param, which can be used to indicate if the user should be redirected to the current page.
 -   Removes `getRefreshURLDomain` function from session recipe.
 -   Removes `SIGN_OUT` pre API hook as it's used within the supertokens-website SDK.
+-   Change to type of `preAPIHook` function
 
 ### Fixed: